#379 refactor admin

+++ aaran/Sundar/Database/Migrations/2024_03_01_001005_create_share_trades_table.php
@@ -6,29 +6,23 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('share_trades', function (Blueprint $table) {
             $table->id();
             $table->string('vdate')->nullable();
             $table->decimal('opening_balance')->nullable();
-            $table->decimal('deposit',10,3)->nullable();
-            $table->decimal('loosed',10,3)->nullable();
-            $table->decimal('withdraw',10,3)->nullable();
-            $table->decimal('charges',10,3)->nullable();
-            $table->decimal('balance',10,3)->nullable();
+            $table->decimal('deposit',13,2)->nullable();
+            $table->decimal('loosed',13,2)->nullable();
+            $table->decimal('withdraw',13,2)->nullable();
+            $table->decimal('charges',13,2)->nullable();
+            $table->decimal('balance',13,2)->nullable();
             $table->string('remarks')->nullable();
-            $table->decimal('active_id',3)->nullable();
+            $table->smallInteger('active_id')->nullable();
 
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('share_trades');
